Refactor messaging view into modular screens
