+++ lib/modules/messaging/views/widgets/scrollable_placeholder.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class ScrollablePlaceholder extends StatelessWidget {
+  const ScrollablePlaceholder({super.key, required this.child});
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        return ListView(
+          physics: const BouncingScrollPhysics(
+            parent: AlwaysScrollableScrollPhysics(),
+          ),
+          children: [
+            ConstrainedBox(
+              constraints: BoxConstraints(minHeight: constraints.maxHeight),
+              child: Center(child: child),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
